requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
fake-useragent==0.1.11
idna==2.10
lxml==4.6.1
pycparser==2.20
pyhocon==0.3.57
pyparsing==2.4.7
python-telegram-bot==13.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
